scripts/IndexTools/requirements.txt
typesense
python-dotenv
markdown
python-frontmatter